+++ NEWS.md
@@ -1,6 +1,8 @@
 # gfilmm 0.1.0.9000
 
 * Fixed a mistake in the C++ code.
+* The main function `gfilmm` has now an option `long`, to run the algorithm with 
+long doubles instead of doubles.
 
 # gfilmm 0.1.0